Fixed long wait bug when scaling up a cluster with the -g flag.
When using the -g flag to scale up a cluster, the wait for all pods to
be up would never succeed and always reach the timeout when there is
more than one Marathon application with the same cluster name. Note that
this bug was actually not leading to an error since the _spin() would
return eventually (even if the condition is not met), but it would wait
for the duration of the specified timeout for nothing.